Add emoji categories support (program level)
Categories do not affect the display of emojis in the GUI yet. As a category, the name of the parent folder in which the emoji is located is taken. Folders between the parent and root "emoji" folder are ignored. If an emoji is added to the "emoji" root folder, the emoji is assigned the "other" category by default. Both Emoji and EmojiHandler stores information about the category